feat: rename classes.ts
